Factor out online_reps and gap_cache (#2070)
* Factor out online_reps and gap_cache

* Formatting not dealt with by clang-format-all
