+++ nano/node/CMakeLists.txt
@@ -32,6 +32,8 @@ add_library (node
 	diagnosticsconfig.cpp
 	election.hpp
 	election.cpp
+	gap_cache.hpp
+	gap_cache.cpp
 	ipc.hpp
 	ipc.cpp
 	ipcconfig.hpp
@@ -56,6 +58,8 @@ add_library (node
 	node_rpc_config.cpp
 	node.hpp
 	node.cpp
+	online_reps.hpp
+	online_reps.cpp
 	openclconfig.hpp
 	openclconfig.cpp
 	openclwork.hpp
