+++ nano/node/node.cpp
@@ -650,121 +650,6 @@ std::unique_ptr<seq_con_info_component> collect_seq_con_info (node & node, const
 }
 }
 
-nano::gap_cache::gap_cache (nano::node & node_a) :
-node (node_a)
-{
-}
-
-void nano::gap_cache::add (nano::transaction const & transaction_a, nano::block_hash const & hash_a, std::chrono::steady_clock::time_point time_point_a)
-{
-	std::lock_guard<std::mutex> lock (mutex);
-	auto existing (blocks.get<1> ().find (hash_a));
-	if (existing != blocks.get<1> ().end ())
-	{
-		blocks.get<1> ().modify (existing, [time_point_a](nano::gap_information & info) {
-			info.arrival = time_point_a;
-		});
-	}
-	else
-	{
-		blocks.insert ({ time_point_a, hash_a, std::vector<nano::account> () });
-		if (blocks.size () > max)
-		{
-			blocks.get<0> ().erase (blocks.get<0> ().begin ());
-		}
-	}
-}
-
-void nano::gap_cache::vote (std::shared_ptr<nano::vote> vote_a)
-{
-	std::lock_guard<std::mutex> lock (mutex);
-	auto transaction (node.store.tx_begin_read ());
-	for (auto hash : *vote_a)
-	{
-		auto existing (blocks.get<1> ().find (hash));
-		if (existing != blocks.get<1> ().end ())
-		{
-			auto is_new (false);
-			blocks.get<1> ().modify (existing, [&is_new, &vote_a](nano::gap_information & info) {
-				auto it = std::find (info.voters.begin (), info.voters.end (), vote_a->account);
-				is_new = (it == info.voters.end ());
-				if (is_new)
-				{
-					info.voters.push_back (vote_a->account);
-				}
-			});
-
-			if (is_new)
-			{
-				uint128_t tally;
-				for (auto & voter : existing->voters)
-				{
-					tally += node.ledger.weight (transaction, voter);
-				}
-				bool start_bootstrap (false);
-				if (!node.flags.disable_lazy_bootstrap)
-				{
-					if (tally >= node.config.online_weight_minimum.number ())
-					{
-						start_bootstrap = true;
-					}
-				}
-				else if (!node.flags.disable_legacy_bootstrap && tally > bootstrap_threshold (transaction))
-				{
-					start_bootstrap = true;
-				}
-				if (start_bootstrap)
-				{
-					auto node_l (node.shared ());
-					auto now (std::chrono::steady_clock::now ());
-					node.alarm.add (node_l->network_params.network.is_test_network () ? now + std::chrono::milliseconds (5) : now + std::chrono::seconds (5), [node_l, hash]() {
-						auto transaction (node_l->store.tx_begin_read ());
-						if (!node_l->store.block_exists (transaction, hash))
-						{
-							if (!node_l->bootstrap_initiator.in_progress ())
-							{
-								node_l->logger.try_log (boost::str (boost::format ("Missing block %1% which has enough votes to warrant lazy bootstrapping it") % hash.to_string ()));
-							}
-							if (!node_l->flags.disable_lazy_bootstrap)
-							{
-								node_l->bootstrap_initiator.bootstrap_lazy (hash);
-							}
-							else if (!node_l->flags.disable_legacy_bootstrap)
-							{
-								node_l->bootstrap_initiator.bootstrap ();
-							}
-						}
-					});
-				}
-			}
-		}
-	}
-}
-
-nano::uint128_t nano::gap_cache::bootstrap_threshold (nano::transaction const & transaction_a)
-{
-	auto result ((node.online_reps.online_stake () / 256) * node.config.bootstrap_fraction_numerator);
-	return result;
-}
-
-size_t nano::gap_cache::size ()
-{
-	std::lock_guard<std::mutex> lock (mutex);
-	return blocks.size ();
-}
-
-namespace nano
-{
-std::unique_ptr<seq_con_info_component> collect_seq_con_info (gap_cache & gap_cache, const std::string & name)
-{
-	auto count = gap_cache.size ();
-	auto sizeof_element = sizeof (decltype (gap_cache.blocks)::value_type);
-	auto composite = std::make_unique<seq_con_info_composite> (name);
-	composite->add_component (std::make_unique<seq_con_info_leaf> (seq_con_info{ "blocks", count, sizeof_element }));
-	return composite;
-}
-}
-
 void nano::node::process_active (std::shared_ptr<nano::block> incoming)
 {
 	block_arrival.add (incoming->hash ());
@@ -1640,101 +1525,6 @@ std::unique_ptr<seq_con_info_component> collect_seq_con_info (block_arrival & bl
 	}
 
 	auto sizeof_element = sizeof (decltype (block_arrival.arrival)::value_type);
-	auto composite = std::make_unique<seq_con_info_composite> (name);
-	composite->add_component (std::make_unique<seq_con_info_leaf> (seq_con_info{ "arrival", count, sizeof_element }));
-	return composite;
-}
-}
-
-nano::online_reps::online_reps (nano::node & node_a, nano::uint128_t minimum_a) :
-node (node_a),
-minimum (minimum_a)
-{
-	auto transaction (node.ledger.store.tx_begin_read ());
-	online = trend (transaction);
-}
-
-void nano::online_reps::observe (nano::account const & rep_a)
-{
-	auto transaction (node.ledger.store.tx_begin_read ());
-	if (node.ledger.weight (transaction, rep_a) > 0)
-	{
-		std::lock_guard<std::mutex> lock (mutex);
-		reps.insert (rep_a);
-	}
-}
-
-void nano::online_reps::sample ()
-{
-	auto transaction (node.ledger.store.tx_begin_write ());
-	// Discard oldest entries
-	while (node.ledger.store.online_weight_count (transaction) >= node.network_params.node.max_weight_samples)
-	{
-		auto oldest (node.ledger.store.online_weight_begin (transaction));
-		assert (oldest != node.ledger.store.online_weight_end ());
-		node.ledger.store.online_weight_del (transaction, oldest->first);
-	}
-	// Calculate current active rep weight
-	nano::uint128_t current;
-	std::unordered_set<nano::account> reps_copy;
-	{
-		std::lock_guard<std::mutex> lock (mutex);
-		reps_copy.swap (reps);
-	}
-	for (auto & i : reps_copy)
-	{
-		current += node.ledger.weight (transaction, i);
-	}
-	node.ledger.store.online_weight_put (transaction, std::chrono::system_clock::now ().time_since_epoch ().count (), current);
-	auto trend_l (trend (transaction));
-	std::lock_guard<std::mutex> lock (mutex);
-	online = trend_l;
-}
-
-nano::uint128_t nano::online_reps::trend (nano::transaction & transaction_a)
-{
-	std::vector<nano::uint128_t> items;
-	items.reserve (node.network_params.node.max_weight_samples + 1);
-	items.push_back (minimum);
-	for (auto i (node.ledger.store.online_weight_begin (transaction_a)), n (node.ledger.store.online_weight_end ()); i != n; ++i)
-	{
-		items.push_back (i->second.number ());
-	}
-
-	// Pick median value for our target vote weight
-	auto median_idx = items.size () / 2;
-	nth_element (items.begin (), items.begin () + median_idx, items.end ());
-	return nano::uint128_t{ items[median_idx] };
-}
-
-nano::uint128_t nano::online_reps::online_stake () const
-{
-	std::lock_guard<std::mutex> lock (mutex);
-	return std::max (online, minimum);
-}
-
-std::vector<nano::account> nano::online_reps::list ()
-{
-	std::vector<nano::account> result;
-	std::lock_guard<std::mutex> lock (mutex);
-	for (auto & i : reps)
-	{
-		result.push_back (i);
-	}
-	return result;
-}
-
-namespace nano
-{
-std::unique_ptr<seq_con_info_component> collect_seq_con_info (online_reps & online_reps, const std::string & name)
-{
-	size_t count = 0;
-	{
-		std::lock_guard<std::mutex> guard (online_reps.mutex);
-		count = online_reps.reps.size ();
-	}
-
-	auto sizeof_element = sizeof (decltype (online_reps.reps)::value_type);
 	auto composite = std::make_unique<seq_con_info_composite> (name);
 	composite->add_component (std::make_unique<seq_con_info_leaf> (seq_con_info{ "arrival", count, sizeof_element }));
 	return composite;
