+++ nano/node/online_reps.hpp
@@ -1,0 +1,42 @@
+#pragma once
+
+#include <nano/lib/numbers.hpp>
+#include <nano/lib/utility.hpp>
+
+#include <memory>
+#include <unordered_set>
+#include <vector>
+
+namespace nano
+{
+class node;
+class transaction;
+
+/** Track online representatives and trend online weight */
+class online_reps final
+{
+public:
+	online_reps (nano::node &, nano::uint128_t);
+
+	/** Add voting account \p rep_account to the set of online representatives */
+	void observe (nano::account const & rep_account);
+	/** Called periodically to sample online weight */
+	void sample ();
+	/** Returns the trended online stake, but never less than configured minimum */
+	nano::uint128_t online_stake () const;
+	/** List of online representatives */
+	std::vector<nano::account> list ();
+
+private:
+	nano::uint128_t trend (nano::transaction &);
+	mutable std::mutex mutex;
+	nano::node & node;
+	std::unordered_set<nano::account> reps;
+	nano::uint128_t online;
+	nano::uint128_t minimum;
+
+	friend std::unique_ptr<seq_con_info_component> collect_seq_con_info (online_reps & online_reps, const std::string & name);
+};
+
+std::unique_ptr<seq_con_info_component> collect_seq_con_info (online_reps & online_reps, const std::string & name);
+}
